Cast return value of SleepDeferrer to int

+++ src/SleepDeferrer.php
@@ -6,7 +6,7 @@ class SleepDeferrer implements Deferrer
 {
     public function getCurrentTime(): int
     {
-        return round(microtime(true) * 1000);
+        return (int)round(microtime(true) * 1000);
     }
 
     public function sleep(int $milliseconds)
